more work on library
Most work here is adding support for more models. Also added the ability to have a wait loop callback for while the ePaper library is waiting for the device to reurn to its ready state. This can be used to allow other libraries to do work. For example, you could add a callback that itself calls the ESP8266WebServer's handleClient() call so that the web server can process requests while your app is waiting on the ePaper device to be ready.